--- ansible.cfg
@@ -1,10 +1,0 @@
-[defaults]
-remote_user =ec2-user
-inventory      = ./ansible/hosts
-sudo_user      = ec2-user
-host_key_checking = False
-
-[privilege_escalation]
-become=True
-become_method=sudo
-become_user=ec2-user
