+++ model/discover.py
@@ -1,0 +1,12 @@
+from mongoengine import *
+
+class Post(Document):
+    host = StringField(required=True, max_length=200, unique=True)
+    ip = StringField(required=True, unique=True)
+    subnet = StringField(required=True, max_length=50)
+    network = StringField(required=True, max_length=50)
+    ports = ListField()
+    cores = StringField(max_length=2)
+    cpu_model = StringField(required=True, max_length=150)
+    ram = StringField(required=True, max_length=50)
+    disk = StringField(required=True, max_length=50)
