+++ pkg/aws/create_ami.py
