+++ utils/log_reader.py
@@ -1,0 +1,5 @@
+def read_logs():
+    l = ''
+    f = open('./ansible/log.txt','r')
+    l = f.read()
+    return l
